# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "joinedcontext Portal API",
    "description": "Administrative and platform management REST API for joinedcontext Portal",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project}/ops": {
      "get": {
        "tags": [
          "ops"
        ],
        "operationId": "list_ops",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of operations available to caller",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/ops/{name}": {
      "post": {
        "tags": [
          "ops"
        ],
        "operationId": "run_op",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Operation name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Operation input parameters",
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "202": {
            "description": "Operation created a change proposal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Operation or project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input according to schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Lane": {
        "type": "string",
        "description": "Change risk classification lane (CC-63).",
        "enum": [
          "green",
          "yellow",
          "red"
        ]
      },
...
      "OperationAnnotations": {
        "type": "object",
        "required": [
          "readOnlyHint",
          "destructiveHint",
          "idempotentHint"
        ],
        "properties": {
          "destructiveHint": {
            "type": "boolean"
          },
          "idempotentHint": {
            "type": "boolean"
          },
          "readOnlyHint": {
            "type": "boolean"
          }
        }
      },
      "OperationSummary": {
        "type": "object",
        "required": [
          "name",
          "title",
          "description",
          "inputSchema",
          "outputSchema",
          "annotations",
          "lane"
        ],
        "properties": {
          "annotations": {
            "$ref": "#/components/schemas/OperationAnnotations"
          },
          "description": {
            "type": "string"
          },
          "inputSchema": {
            "type": "object"
          },
          "lane": {
            "$ref": "#/components/schemas/Lane"
          },
          "name": {
            "type": "string"
          },
          "outputSchema": {
            "type": "object"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details representation.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every violation of one request, when there is more than one thing to say. A form marks\nall its bad fields from this in one pass instead of one round trip per mistake (CC-24)."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}